pyproject.toml
[project]
name = "ubq"
version = "0.1.0"
description = "Ubuntu Query"

authors = [
	{ name = "Lena Voytek", email = "lena@voytek.dev" },
]

readme = "README.md"
dependencies = [
    "launchpadlib>=2.1.0",
    "lazr-restfulclient>=0.14.6",
]

[dependency-groups]
format = [
    "black>=26.3.1",
    "isort>=8.0.1",
]
lint = [
    "codespell>=2.4.2",
    "mypy>=1.20.2",
    "pylint>=4.0.5",
    "pyproject-flake8>=7.0.0",
]

[tool.black]
line-length = 99

[tool.flake8]
max-line-length = 99
max-doc-length = 99

[tool.isort]
profile = "black"
line_length = 99

[tool.pylint]
max-line-length = 99

[tool.pylint."messages control"]
disable = [
    "R0903", # Too few public methods
    "R0902", # Too many instance attributes
    "C0415", # Import outside toplevel
]
